requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiolimiter==1.1.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
chatgpt_md_converter==0.3.11
distro==1.9.0
dnspython==2.8.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jiter==0.12.0
magic-filter==1.0.12
multidict==6.7.0
openai==2.8.1
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
python-pptx
pypdf
python-docx

